requirements.txt
aiomysql==0.2.0
asgiref==3.8.1
asyncio==3.4.3
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.110
botocore==1.34.110
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
filelock==3.14.0
Flask==3.0.3
gdown==5.2.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
numpy==1.26.4
pycparser==2.22
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.1
s3transfer==0.10.1
six==1.16.0
soupsieve==2.5
tqdm==4.66.4
urllib3==2.2.1
Werkzeug==3.0.3
